Address PR review comments for ED-23495
- Rename session storage key from 'last-converted-class-generated-name' to 'pending-class-rename-id' for better generality
- Extract DuplicateClassMenuItem to a dedicated file (duplicate-class-menu-item.tsx) and move onClick logic to a named handleDuplicate method
- Revert unintended changes to global-classes-list.tsx

Made-with: Cursor
