+++ packages/packages/core/editor-editing-panel/src/components/css-classes/css-class-item.tsx
@@ -54,7 +54,7 @@ export function CssClassItem( props: CssClassItemProps ) {
 	const { userCan } = useUserStylesCapability();
 
 	const [ convertedFromLocalId, , clearConvertedFromLocalId ] = useSessionStorage(
-		'last-converted-class-generated-name',
+		'pending-class-rename-id',
 		'app'
 	);
